+++ lib/truffle.rb
@@ -2,6 +2,7 @@
 
 require_relative "truffle/version"
 require_relative "truffle/content"
+require_relative "truffle/stop_reason"
 require_relative "truffle/message"
 require_relative "truffle/response"
 require_relative "truffle/tool"